scripts/firebolt_setup.py
from firebolt.db import connect
import pandas as pd
from pyiceberg.catalog import load_catalog
from pyiceberg.table import Table

# Firebolt connection parameters
# Replace these with your actual Firebolt credentials
FIREBOLT_ACCOUNT = "your-account-name"
FIREBOLT_USERNAME = "your-username"
FIREBOLT_PASSWORD = "your-password"
FIREBOLT_DATABASE = "your-database"
FIREBOLT_ENGINE = "your-engine-name"

def create_firebolt_connection():
    """Create connection to Firebolt"""
    connection = connect(
        account_name=FIREBOLT_ACCOUNT,
        username=FIREBOLT_USERNAME,
        password=FIREBOLT_PASSWORD,
        database=FIREBOLT_DATABASE,
        engine_name=FIREBOLT_ENGINE
    )
    return connection

def setup_iceberg_catalog_firebolt():
    """Setup Iceberg catalog pointing to S3/MinIO"""
    catalog = load_catalog(
        "firebolt_catalog",
        **{
            "type": "rest",
            "uri": "http://localhost:8181",  # Iceberg REST catalog endpoint
            "s3.endpoint": "http://minio:9000",
            "s3.access-key-id": "minioadmin",
            "s3.secret-access-key": "minioadmin",
            "warehouse": "s3://iceberg-warehouse/"
        }
    )
    return catalog

if __name__ == "__main__":
    print("Firebolt setup utilities")
    print("Configure your Firebolt credentials in this file before use.")

