requirements.txt
pyspark==3.5.0
pyiceberg==0.7.1
minio==7.2.0
boto3==1.34.0
pandas==2.1.4
pyarrow==14.0.1
jupyterlab==4.0.11
requests==2.31.0
firebolt-sdk>=1.0.0
